test: Car 테스트에서 RaceJudge 테스트 분리
